+++ src/Sentry/Platforms/Cocoa/SentryOptions.cs
@@ -182,7 +182,7 @@ public partial class SentryOptions
         /// Gets or sets a value that indicates if tracing features are enabled on the embedded Cocoa SDK.
         /// The default value is <c>false</c> (disabled).
         /// </summary>
-        public bool EnableCocoaSdkTracing { get; set; } = false;
+        public bool EnableTracing { get; set; } = false;
 
         internal List<string>? InAppExcludes { get; private set; }
         internal List<string>? InAppIncludes { get; private set; }